controller_db: Upgrade Traffic Stats to 64-bits
In TR-181, Traffic Stats are defined as 64-bits values, while they are
defined as 32-bits values in EasyMesh.
When a rollover is detected in the value reported in EasyMesh message,
increase the TR-181 version to use the full 64-bits range.

+++ controller/src/beerocks/master/db/db.h
@@ -249,8 +249,8 @@ public:
     typedef struct {
         uint64_t m_byte_sent            = 0;
         uint64_t m_byte_received        = 0;
-        uint64_t m_packets_sent         = 0;
-        uint64_t m_packets_received     = 0;
+        uint32_t m_packets_sent         = 0;
+        uint32_t m_packets_received     = 0;
         uint32_t m_tx_packets_error     = 0;
         uint32_t m_rx_packets_error     = 0;
         uint32_t m_retransmission_count = 0;
